Move receiver functions to the same place
